Pick-up boost and cppunit from wherever Homebrew installed them.
On macOS, Intel versus Apple silicon has created variability as far
as Homebrew is concerned. On the former, Homebrew formulae are installed
in /usr/local and on the latter, in /opt/homebrew. There's no way to know
a priori on which architecture the job will run. Consequently, we need to
handle both.